Move tensorflow_silence into separate file
